+++ common/datablocks/tests/it/kernels/data_block_group_by_hash.rs
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_datablocks::*;
 use common_datavalues::prelude::*;
 use common_exception::Result;
-
-use crate::*;
 
 #[test]
 fn test_data_block_group_by_hash() -> Result<()> {
