+++ common/datablocks/tests/it/kernels/data_block_slice.rs
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_datablocks::*;
 use common_datavalues::prelude::*;
 use common_exception::Result;
-
-use crate::*;
 
 #[test]
 fn test_data_block_slice() -> Result<()> {
@@ -43,6 +42,6 @@ fn test_data_block_slice() -> Result<()> {
         "| 5 | 5 |",
         "+---+---+",
     ];
-    crate::assert_blocks_eq(expected, &sliced);
+    common_datablocks::assert_blocks_eq(expected, &sliced);
     Ok(())
 }
