+++ common/datablocks/tests/it/kernels/mod.rs
@@ -20,6 +20,3 @@ mod data_block_scatter;
 mod data_block_slice;
 mod data_block_sort;
 mod data_block_take;
-
-pub use data_block_group_by_hash::*;
-pub use data_block_sort::SortColumnDescription;
